Move out command types to avoid circular dependency and simplify index to the public API.
